+++ src/my_utilities/convenience/__init__.py
@@ -1,5 +1,8 @@
 from ._convenience import *
-from ._formaters import PrettyPrinter
+from ._formaters import (
+    PrettyPrinter, 
+    TextWrap,
+)
 from ._containers import *
 from ._linkage_tree_parser import LinkageTreeParser
 from ._restricted_logger import Logger